Added Makefile help command snippet to README
The README file has been updated with a new section that provides a code snippet for creating a Makefile with a help command. This addition includes an explanation of how the `awk` command is used to parse the Makefile and display commands along with their descriptions. An example output of this process is also provided for clarity.