Added sqlc configuration file
A new sqlc.yaml file has been added to configure SQLC for the project. This includes setting up the schema, queries, and engine details. The Go package is specified along with output directory and SQL package. JSON tags are set to be emitted, as well as interfaces and empty slices. Overrides have also been defined for specific database types to corresponding Go types.